[SW-30] - вынесен в отдельный файл VisualEffectBadgeView
